Broke comment_parse into all of its grouped functions
