+++ tipbot/backend/transaction.py
