+++ tipbot/donate.py
@@ -1,0 +1,58 @@
+import re
+from decimal import Decimal
+
+import helper
+from helper import general_fund_address, format_decimal, signature, get_xmr_val
+from logger import tipper_logger
+from tipbot.backend.safe_wallet import safe_wallet
+from tipbot.backend.transaction import generate_transaction
+
+
+def parse_donate_amount(subject, senderBalance):
+    """
+    Parses the amount a user wishes to donate to the CSS based on their message.
+
+    :param subject: Subject line in form "Donate xyz XMR" or "Donate xyz% of my balance"
+    :param senderBalance: Their current balance, used to calculate when sending a percentage
+    :return: Final amount in XMR they wish to donate
+    """
+
+    # "Donate xyz XMR"
+    m = re.search('donate ([\\d\\.]+)( )?(m)?xmr', subject.lower())
+    if m:
+        return str(Decimal(m.group(1))/1000) if m.group(m.lastindex) == "m" else m.group(1)
+
+    # "Donate xyz% of my balance"
+    m = re.search('donate ([\\d\\.]+)% of my balance', subject.lower())
+    if m:
+        return str(float(m.group(1))*float(senderBalance)/100)
+
+    # "Donate xyz$
+    m = re.search('donate (\\$)?(?P<dollar_amt>[\\d\\.]+)(\\$)?', str(subject).lower())
+    if m:
+        return str(get_xmr_val(m.group("dollar_amt")))
+
+
+def handle_donation(author, subject):
+    """
+    Allows Reddit users to donate a portion of their balance directly to the CCS
+    CCS can be seen at: https://ccs.getmonero.org/
+
+    :param author: Reddit account to withdraw from
+    :param subject: Subject line of the message, telling how much to withdraw
+    """
+
+    sender_rpc_n_wallet = safe_wallet(port=helper.ports.donation_sender_port, wallet_name=author.name.lower(), password=helper.bot_handler.password)
+
+    amount = Decimal(parse_donate_amount(subject, sender_rpc_n_wallet.wallet.balance()))
+
+    try:
+        generate_transaction(senderWallet=sender_rpc_n_wallet.wallet, recipientAddress=general_fund_address, amount=amount, splitSize=1)
+        helper.bot_handler.reddit.redditor(author.name).message(subject="Your donation to the General Dev Fund", message=f'Thank you for donating {format_decimal(amount)} of your XMR balance to the CCS!\n\nYou will soon have your total donations broadcasted to the wiki :) {signature}')
+        helper.bot_handler.reddit.redditor("OsrsNeedsF2P").message(subject=f'{author.name} donated {amount} to the CCS!', message="Update table here: https://old.reddit.com/r/MoneroTipsBot/wiki/index#wiki_donating_to_the_ccs")
+        tipper_logger.log(f'{author.name} donated {format_decimal(amount)} to the CCS.')
+    except Exception as e:
+        helper.bot_handler.reddit.redditor(author.name).message(subject="Your donation to the CCS failed", message=f'Please send the following to /u/OsrsNeedsF2P:\n\n' + str(e) + signature)
+        tipper_logger.log("Caught an error during a donation to CCS: " + str(e))
+
+    sender_rpc_n_wallet.kill_rpc()
